[JIT] Move `to_backend` TestBackend to test/cpp/jit
Summary:
This commit moves the `TestBackend` that is currently in
`torch/csrc/jit/backends` to `torch/test/cpp/jit`. To do so, it creates
another Python C extension module in `test/tech/cpp/jit`, registers
the `to_test_backend` function in that module, and uses that function
in `test/jit/test_backends.py` instead of
`torch._C._jit_to_test_backend`. In addition to maintaining separation
between source and test code, this also reflects the manner in which
this extension point is intended to be used; backend writers should
register the function passed back by `generateToBackendFn` in C
extension module that they own and export one or more Python functions
that wraps it.

Test Plan:
`python test/test_jit.py TestBackends`
